MAde The input/ validation loop
Built the input loop. This prints a prompt asking the user to input a number and stores it in a register. The number is then validated to ensure it’s at least 25. If not, the program will display an error message and prompt again.
Used input() in my python code, but here im using syscall to prompt the user and read whatever is inputed(stored in the register).